pyproject.toml
[project]
name = "test-tg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.27",
]
